fix: util 폴더 이동
